long 변환 추가작업

+++ src/main/java/com/meta/ponkids/domain/user/repository/UserChldrnRepository.java
@@ -20,7 +20,7 @@ import java.util.List;
  * -----------------------------------------------------------
  * 2023-11-19        jjeoV             최초 생성
  */
-public interface UserChldrnRepository extends JpaRepository<UserChldrn, Integer> {
+public interface UserChldrnRepository extends JpaRepository<UserChldrn, Long> {
     
     List<UserChldrn> findByUserSn( Long userSn );
